git: implement restore last branch functionality
The squash command was destroying the history of the last branch you
were in, so after you ran it "git co -" would fail. This is because the
last checked out branch was actually a temporary branch that was created
and subsequently deleted to facilitate the squashing.

This commit adds a function for storing what that last branch was and
then restoring it after the squash